requirements.txt
httpx~=0.27.0
uvicorn~=0.22.0
fastapi~=0.109.2
pydantic~=2.5.2
python-dotenv~=1.0.1
deepeval
langchain-core
langchain-community
langchain-text-splitters